xrRender: GL, DX9 and DX10/11 device classes separated
