+++ src/Layers/xrRenderPC_GL/stdafx.h
@@ -26,7 +26,7 @@
 
 #include "xrParticles/psystem.h"
 
-#include "Layers/xrRender/HW.h"
+#include "Layers/xrRenderGL/glHW.h"
 #include "Layers/xrRender/Shader.h"
 #include "Layers/xrRender/R_Backend.h"
 #include "Layers/xrRender/R_Backend_Runtime.h"
