+++ src/Layers/xrRenderPC_R3/stdafx.h
@@ -24,7 +24,7 @@
 #define RENDER R_R3
 
 #include "xrParticles/psystem.h"
-#include "Layers/xrRender/HW.h"
+#include "Layers/xrRenderDX10/dx10HW.h"
 #include "Layers/xrRender/Shader.h"
 #include "Layers/xrRender/R_Backend.h"
 #include "Layers/xrRender/R_Backend_Runtime.h"
